(PUP-6675) Ensure expected encoding on output from pipe
When reading from the pipe in blocks, we allow binary format. This is
primarily to avoid issues if we read only some bytes of a multi-byte
UTF-8 character. At the end, we assign the default external encoding and
convert to internal encoding, as we would expect to get from reading a
file (preserving the prior behavior).